+++ Frontend/src/routes/+page.server.ts
@@ -62,10 +62,10 @@ export const load: PageServerLoad = () => {
   };
 };
 
-export const actions = {
+export const actions: Actions = {
   default: async ({request}) => {
     const data = await request.formData();
     const email = data.get('email') as string;
     return await fetchApi('/join', 'POST', {email});
   },
-} satisfies Actions;
+};