+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.00] - 2024-11-10
+
+### Added
+
+- Finished legal and support pages
+- Add working contact form
+- Add ability to send email
+
+### Changed
+
+- Moved custom tailwind class to tailwind.config.js
+
 ## [0.1.07] - 2024-11-09
 
 ### Added